set length type to float

+++ napari/components/overlays/scale_bar.py
@@ -35,7 +35,7 @@ class ScaleBarOverlay(CanvasOverlay):
     unit : Optional[str]
         Unit to be used by the scale bar. The value can be set
         to `None` to display no units.
-    length : Optional[int]
+    length : Optional[float]
         Fixed length of the scale bar in physical units. If set to `None`,
         it is determined automatically based on zoom level.
     position : CanvasPosition
@@ -57,4 +57,4 @@ class ScaleBarOverlay(CanvasOverlay):
         default_factory=lambda: ColorValue([0, 0, 0, 0.6])
     )
     unit: Optional[str] = None
-    length: Optional[int] = None
+    length: Optional[float] = None
